backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo==4.6.1
motor==3.3.2  # 异步MongoDB驱动
redis==5.0.1

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1

# 图片处理
Pillow==10.2.0

# AI服务
replicate==0.22.0

# 对象存储
oss2==2.18.4

# 任务队列
celery==5.3.4
redis==5.0.1

# 工具库
python-dotenv==1.0.0
tenacity==8.2.3  # 重试
loguru==0.7.2  # 日志

# 日期时间
python-dateutil==2.8.2

# Sentry错误追踪
sentry-sdk[fastapi]==1.39.2

# 限流
slowapi==0.1.9

# 测试相关（可选，建议放到requirements-dev.txt）
# pytest==7.4.4
# pytest-asyncio==0.23.3
# pytest-cov==4.1.0
# httpx==0.26.0

# 地理位置
# geoalchemy2==0.14.3
# shapely==2.0.2
